# static/openapi.json
{"swagger": "2.0", "info": {"title": "Swagger Students", "description": "This is a home task for the Python workshop", "termsOfService": "http://swagger.io/terms/", "contact": {"email": "apiteam@swagger.io"}, "license": {"name": "Apache 2.0", "url": "http://www.apache.org/licenses/LICENSE-2.0.html"}, "version": "1.1.0"}, "servers": [{"url": "http://192.168.0.33:5000/"}], "tags": [{"name": "student", "description": "Everything about your Students"}], "paths": {"/api/students/add": {"post": {"tags": ["student"], "produces": ["application/json"], "responses": {"201": {"description": "the operation was successful"}}, "parameters": [{"in": "body", "name": "body", "description": "Student object", "required": true, "schema": {"$ref": "#/definitions/Student"}}]}}, "/api/students": {"get": {"tags": ["student"], "produces": ["application/json"], "responses": {"200": {"description": "the operation was successful"}}}}, "/api": {"get": {"tags": ["student"], "produces": ["application/json"], "responses": {"200": {"description": "the operation was successful"}}}}, "/api/students/change/{id}": {"put": {"tags": ["student"], "produces": ["application/json"], "responses": {"200": {"description": "the operation was successful"}}, "parameters": [{"in": "body", "name": "body", "description": "Student object", "required": true, "schema": {"$ref": "#/definitions/Student"}}, {"name": "id", "in": "path", "description": "ID of the Student", "required": true, "type": "integer"}]}}, "/api/students/modify/{id}": {"patch": {"tags": ["student"], "produces": ["application/json"], "responses": {"200": {"description": "the operation was successful"}}, "parameters": [{"in": "body", "name": "body", "description": "Student object", "required": true, "schema": {"$ref": "#/definitions/Student"}}, {"name": "id", "in": "path", "description": "ID of the Student", "required": true, "type": "integer"}]}}, "/api/students/delete/{id}": {"delete": {"tags": ["student"], "produces": ["application/json"], "responses": {"204": {"description": "the operation was successful"}}, "parameters": [{"name": "id", "in": "path", "description": "ID of the Student", "required": true, "type": "integer"}]}}, "/api/students/get/{id}": {"get": {"tags": ["student"], "produces": ["application/json"], "responses": {"200": {"description": "the operation was successful"}}, "parameters": [{"name": "id", "in": "path", "description": "ID of the Student", "required": true, "type": "integer"}]}}}, "definitions": {"Student": {"type": "object", "properties": {"name": {"type": "string"}, "age": {"type": "integer"}, "email": {"type": "string"}, "cellphone": {"type": "string"}}, "xml": {"name": "Student"}}}}